Chain create/update of eb environemtns
From the comminication with AWS (after UPDATE_ROLLBACK_FAILED):

Your CloudFormation stack contains several Beanstalk environments,
and Beanstalk tried to update them all in parallel. This generated
a large volume of parallel calls to the EC2 APIs, resulting in
the Request Limit Exceeded error.

CloudFormation then tried to roll back the changes, i.e. revert
to the previous AMI. But several of the environments couldn't be
reverted as the first change was still pending.

This resulted in the UPDATE_ROLLBACK_FAILED status.

I'll ask our CloudFormation team to move the stack to
an UPDATE_COMPLETE state, but I'll note that this may leave some
of the Beanstalk environments running the new AMI, and some still
running the older one.

In future, to reduce the likelihood of this happening, you should
only update some Beanstalk environments at a time, or use the
CloudFormation DependsOn attribute to serialise some of the updates
(i.e. to ensure one environment has finished updating before
the next one starts).